Refactoring of patching
Patching is now handled on a per-patch basis versus one giant file for patching.  Each patch must pass validation for it to continue.  After a patch has been successfully applied, the version is updated so it will not run again
