+++ configmode.py
@@ -8,7 +8,7 @@ from mapadroid.utils.MappingManager import MappingManager
 from mapadroid.utils.MappingManager import MappingManagerManager
 from mapadroid.utils.logging import initLogging, logger
 from mapadroid.utils.updater import deviceUpdater
-from mapadroid.utils.version import MADVersion
+from mapadroid.patcher import MADVersion
 from mapadroid.utils.walkerArgs import parseArgs
 from mapadroid.utils.data_manager import DataManager
 from mapadroid.websocket.WebsocketServer import WebsocketServer
